Don't delete 3rdparty for Android because it contains cpufeatures library

+++ conanfile.py
@@ -59,7 +59,8 @@ class OpenCVConan(ConanFile):
         tools.get("https://github.com/opencv/opencv_contrib/archive/{}.tar.gz".format(self.version), sha256=sha256)
         os.rename('opencv_contrib-%s' % self.version, 'contrib')
 
-        tools.rmdir(os.path.join(self._source_subfolder, '3rdparty'))
+        if self.settings.os != 'Android':
+            tools.rmdir(os.path.join(self._source_subfolder, '3rdparty'))
 
     def config_options(self):
         if self.settings.os == 'Windows':
